+++ tests/qasm_test.py
@@ -36,7 +36,7 @@ def test_pytket_qir_qasm() -> None:
 )
 def test_pytket_qir_qasm_classical_0(profile: QIRProfile) -> None:
     with open("qasm/test0.qasm") as my_file:
-        circ = circuit_from_qasm_str(my_file.read(), use_clexpr=True)
+        circ = circuit_from_qasm_str(my_file.read())
 
     run_qir_gen_and_check(
         circ,
@@ -55,7 +55,7 @@ def test_pytket_qir_qasm_classical_0(profile: QIRProfile) -> None:
 )
 def test_pytket_qir_qasm_classical_1(profile: QIRProfile) -> None:
     with open("qasm/test1.qasm") as my_file:
-        circ = circuit_from_qasm_str(my_file.read(), use_clexpr=True)
+        circ = circuit_from_qasm_str(my_file.read())
 
     run_qir_gen_and_check(
         circ,